pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask-wtf>=1.2.2",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.32.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-login>=0.6.3",
    "python-docx>=1.2.0",
    "werkzeug>=3.1.3",
    "docx>=0.2.4",
    "sift-stack-py>=0.8.5",
    "telegram>=0.0.1",
    "python-telegram-bot>=22.3",
    "bcrypt>=4.3.0",
    "sqlalchemy>=2.0.43",
]
